Feature: Exclude 'todo' keywords from spellcheck
The configured 'org_todo_keywords' are subject to spellcheck. In
languages other than English, the spellchecker may not consider those
words to be spelled correctly. As a result, the spellchecker incorrectly
and annoyingly flags these words as misspelled. This patch removes the
keywords in the 'org_todo_keywords' from the province of the
spellchecker.